refactor: extendr-options is used by many scripts already
